refactor: transfer responsibility to show Balloon to Portal
